+++ src/Codeception/Module/WPBrowserMethods.php
@@ -7,10 +7,10 @@
 		/**
 		 * Goes to the login page and logs in as the site admin.
 		 *
-		 * @return void
+		 * @return array An array of login credentials and auth cookies.
 		 */
 		public function loginAsAdmin() {
-			$this->loginAs( $this->config['adminUsername'], $this->config['adminPassword'] );
+			return $this->loginAs( $this->config['adminUsername'], $this->config['adminPassword'] );
 		}
 
 		/**
@@ -19,13 +19,48 @@
 		 * @param string $username
 		 * @param string $password
 		 *
-		 * @return void
+		 * @return array An array of login credentials and auth cookies.
 		 */
 		public function loginAs( $username, $password ) {
 			$this->amOnPage( $this->loginUrl );
 			$this->fillField( '#user_login', $username );
 			$this->fillField( '#user_pass', $password );
 			$this->click( '#wp-submit' );
+
+			return [
+				'username'    => $username,
+				'password'    => $password,
+				'authCookie'  => $this->grabWordPressAuthCookie(),
+				'loginCookie' => $this->grabWordPressLoginCookie()
+			];
+		}
+
+		/**
+		 * Returns WordPress default auth cookie if present.
+		 *
+		 * @param null $pattern Optional, overrides the default cookie name.
+		 *
+		 * @return mixed Either a cookie or null.
+		 */
+		public function grabWordPressAuthCookie( $pattern = null ) {
+			$pattern = $pattern ? $pattern : '/^wordpress_[a-z0-9]{32}$/';
+			$cookies = $this->grabCookiesWithPattern( $pattern );
+
+			return empty( $cookies ) ? null : array_pop( $cookies );
+		}
+
+		/**
+		 * Returns WordPress default login cookie if present.
+		 *
+		 * @param null $pattern Optional, overrides the default cookie name.
+		 *
+		 * @return mixed Either a cookie or null.
+		 */
+		public function grabWordPressLoginCookie( $pattern = null ) {
+			$pattern = $pattern ? $pattern : '/^wordpress_logged_in_[a-z0-9]{32}$/';
+			$cookies = $this->grabCookiesWithPattern( $pattern );
+
+			return empty( $cookies ) ? null : array_pop( $cookies );
 		}
 
 		/**
@@ -38,7 +73,7 @@
 		 * @return void
 		 */
 		public function activatePlugin( $pluginSlug ) {
-			$this->click( 'Activate', '#' . $pluginSlug );
+			$this->click( "table.plugins tr[data-slug='{$pluginSlug}'] span.activate > a:first-of-type" );
 		}
 
 		/**
@@ -51,7 +86,7 @@
 		 * @return void
 		 */
 		public function deactivatePlugin( $pluginSlug ) {
-			$this->click( 'Deactivate', '#' . $pluginSlug );
+			$this->click( "table.plugins tr[data-slug='{$pluginSlug}'] span.deactivate > a:first-of-type" );
 		}
 
 		/**
@@ -87,7 +122,7 @@
 		 */
 		public function seePluginDeactivated( $pluginSlug ) {
 			$this->seePluginInstalled( $pluginSlug );
-			$this->seeElement( "#" . $pluginSlug . '.inactive' );
+			$this->seeElement( "table.plugins tr[data-slug='$pluginSlug'].inactive" );
 		}
 
 		/**
@@ -100,7 +135,7 @@
 		 * @return void
 		 */
 		public function seePluginInstalled( $pluginSlug ) {
-			$this->seeElement( '#' . $pluginSlug );
+			$this->seeElement( "table.plugins tr[data-slug='$pluginSlug']" );
 		}
 
 		/**
@@ -114,7 +149,7 @@
 		 */
 		public function seePluginActivated( $pluginSlug ) {
 			$this->seePluginInstalled( $pluginSlug );
-			$this->seeElement( "#" . $pluginSlug . '.active' );
+			$this->seeElement( "table.plugins tr[data-slug='$pluginSlug'].active" );
 		}
 
 		/**
@@ -127,7 +162,7 @@
 		 * @return void
 		 */
 		public function dontSeePluginInstalled( $pluginSlug ) {
-			$this->dontSeeElement( '#' . $pluginSlug );
+			$this->dontSeeElement( "table.plugins tr[data-slug='$pluginSlug']" );
 		}
 
 		/**
@@ -190,31 +225,8 @@
 			return $this->grabCookie( $pattern );
 		}
 
-		/**
-		 * Returns WordPress default login cookie if present.
-		 *
-		 * @param null $pattern Optional, overrides the default cookie name.
-		 *
-		 * @return mixed Either a cookie or null.
-		 */
-		public function grabWordPressLoginCookie( $pattern = null ) {
-			$pattern = $pattern ? $pattern : '/^wordpress_logged_in_[a-z0-9]{32}$/';
-			$cookies = $this->grabCookiesWithPattern( $pattern );
-
-			return empty( $cookies ) ? null : array_pop( $cookies );
-		}
-
-		/**
-		 * Returns WordPress default auth cookie if present.
-		 *
-		 * @param null $pattern Optional, overrides the default cookie name.
-		 *
-		 * @return mixed Either a cookie or null.
-		 */
-		public function grabWordPressAuthCookie( $pattern = null ) {
-			$pattern = $pattern ? $pattern : '/^wordpress_[a-z0-9]{32}$/';
-			$cookies = $this->grabCookiesWithPattern( $pattern );
-
-			return empty( $cookies ) ? null : array_pop( $cookies );
+		public function amOnAdminPage($path)
+		{
+			$this->amOnPage($this->adminUrl . '/' . ltrim($path, '/'));
 		}
 	}