+++ src/includes/install.php
@@ -4,7 +4,6 @@
  *
  * @todo Reuse the init/load code in init.php
  */
-
 error_reporting( E_ALL & ~E_DEPRECATED & ~E_STRICT );
 
 // this replaces the original argument passed in with the `system` function
@@ -15,11 +14,7 @@ $table_prefix = WP_TESTS_TABLE_PREFIX ;
 define( 'WP_INSTALLING', true );
 require_once dirname( __FILE__ ) . '/functions.php';
 
-$_SERVER['SERVER_PROTOCOL'] = 'HTTP/1.1';
-$_SERVER['HTTP_HOST'] = WP_TESTS_DOMAIN;
-$_SERVER['SERVER_NAME'] = WP_TESTS_DOMAIN;
-$_SERVER['REQUEST_METHOD'] = 'GET';
-$_SERVER['REMOTE_ADDR'] = '127.0.0.1';
+tests_reset__SERVER();
 
 $PHP_SELF = $GLOBALS['PHP_SELF'] = $_SERVER['PHP_SELF'] = '/index.php';
 
@@ -33,7 +28,15 @@ global $phpmailer;
 require_once( dirname( __FILE__ ) . '/mock-mailer.php' );
 $phpmailer = new MockPHPMailer();
 
-$wpdb->query( 'SET storage_engine = INNODB' );
+/*
+ * default_storage_engine and storage_engine are the same option, but storage_engine
+ * was deprecated in MySQL (and MariaDB) 5.5.3, and removed in 5.7.
+ */
+if ( version_compare( $wpdb->db_version(), '5.5.3', '>=' ) ) {
+	$wpdb->query( 'SET default_storage_engine = InnoDB' );
+} else {
+	$wpdb->query( 'SET storage_engine = InnoDB' );
+}
 $wpdb->select( DB_NAME, $wpdb->dbh );
 
 echo "Installing..." . PHP_EOL;
@@ -68,6 +71,14 @@ if ( $multisite ) {
 
 	$title = WP_TESTS_TITLE . ' Network';
 	$subdomain_install = false;
+	
+	/**
+	 * If we are on multisite the `blogs` table will not have the main blog row set up due to how the install script ran.
+	 * Let's add the row now.
+	 */
+    /** @var \wpdb $wpdb */
+	global $wpdb;
+	$wpdb->insert($wpdb->blogs, array('site_id' => 1, 'blog_id' => 1, 'domain' => WP_TESTS_DOMAIN, 'path' => '/', 'registered' => current_time('mysql')));
 
 	install_network();
 	populate_network( 1, WP_TESTS_DOMAIN, WP_TESTS_EMAIL, $title, '/', $subdomain_install );